+++ tests/unit/endpoints-factory.spec.ts
@@ -8,7 +8,7 @@ import {
 } from "../../src";
 import { Endpoint } from "../../src/endpoint";
 import { expectType } from "tsd";
-import { AbstractLogger } from "../../src/logger";
+import { ActualLogger } from "../../src/logger-helpers";
 import { makeLoggerMock } from "../../src/testing";
 import { serializeSchemaForTest } from "../helpers";
 import { z } from "zod";
@@ -155,7 +155,7 @@ describe("EndpointsFactory", () => {
           options: {},
           request: requestMock,
           response: responseMock,
-          logger: {} as AbstractLogger,
+          logger: {} as ActualLogger,
         });
         expect(middleware).toHaveBeenCalledTimes(1);
         expect(middleware).toHaveBeenCalledWith(
@@ -188,7 +188,7 @@ describe("EndpointsFactory", () => {
           options: {},
           request: requestMock,
           response: responseMock,
-          logger: {} as AbstractLogger,
+          logger: {} as ActualLogger,
         });
         expect(middleware).toHaveBeenCalledTimes(1);
         expect(middleware).toHaveBeenCalledWith(
@@ -219,7 +219,7 @@ describe("EndpointsFactory", () => {
             options: {},
             request: {} as Request,
             response: {} as Response,
-            logger: {} as AbstractLogger,
+            logger: {} as ActualLogger,
           });
           expect.fail("Should not be here");
         } catch (e) {
@@ -250,7 +250,7 @@ describe("EndpointsFactory", () => {
             options: {},
             request: {} as Request,
             response: {} as Response,
-            logger: {} as AbstractLogger,
+            logger: {} as ActualLogger,
           });
           expect.fail("Should not be here");
         } catch (e) {